build pass on Mac 64-bit, but API test case failed

+++ trunk/mockcpp/include/mockcpp/PageAllocator.h
@@ -34,7 +34,7 @@ struct PageAllocator : public MemAllocator
 
 	void * align(void *addr)
 	{
-		return (void *)(((int)addr) & ~(pageSize() - 1));
+		return (void *)(((size_t)addr) & ~(pageSize() - 1));
 	}
 };
